changes in deteq

+++ src/main/java/com/example/firstcrud/repositories/IDetailEqRepository.java
@@ -3,7 +3,8 @@ package com.example.firstcrud.repositories;
 import com.example.firstcrud.entities.DetailEquipe;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IDetailEqRepository extends CrudRepository<DetailEquipe,Integer> {
+public interface IDetailEqRepository extends CrudRepository<DetailEquipe,Long> {
+
 
 
 }
